Move components to common
