+++ src/Sylius/Bundle/ShopBundle/Resources/config/app/twig_hooks/homepage/index.yaml
@@ -4,12 +4,12 @@ twig_hooks:
             banner:
                 template: '@SyliusShop/homepage/banner.html.twig'
             latest_deals:
-                component: 'sylius_shop:homepage:latest_deals'
+                component: 'sylius_shop:common:latest_deals'
                 props:
                     limit: 4
             new_collection:
                 template: '@SyliusShop/homepage/new_collection.html.twig'
             latest_products:
-                component: 'sylius_shop:homepage:latest_products'
+                component: 'sylius_shop:common:latest_products'
                 props:
                     limit: 8
