+++ src/Sylius/Bundle/ShopBundle/Twig/Component/Common/LatestProductsComponent.php
@@ -11,7 +11,7 @@
 
 declare(strict_types=1);
 
-namespace Sylius\Bundle\ShopBundle\Twig\Component\Homepage;
+namespace Sylius\Bundle\ShopBundle\Twig\Component\Common;
 
 use Sylius\Component\Channel\Context\ChannelContextInterface;
 use Sylius\Component\Core\Model\ChannelInterface;
