+++ src/Sylius/Bundle/ShopBundle/templates/common/latest_products.html.twig
